+++ app/components/application/action/index.ts
@@ -1,5 +1,5 @@
 export * from "./activate/activate-package-dialog.component";
-export * from "./add/application-create-dialog.component";
+export * from "./create/application-create-dialog.component";
 export * from "./delete/delete-application-dialog.component";
 export * from "./delete/delete-application-action";
 export * from "./delete/delete-package-action";
