+++ app/components/application/application.module.ts
@@ -1,23 +1,20 @@
 import { NgModule } from "@angular/core";
 
 import { commonModules } from "app/common";
-import { ActivatePackageDialogComponent } from "./action/activate/activate-package-dialog.component";
-import { ApplicationCreateDialogComponent } from "./action/add/application-create-dialog.component";
-import { DeleteApplicationDialogComponent } from "./action/delete/delete-application-dialog.component";
-import { ApplicationEditDialogComponent } from "./action/edit/application-edit-dialog.component";
-import { ApplicationListComponent } from "./browse/application-list.component";
-import { ApplicationPreviewComponent } from "./browse/preview";
-import { ApplicationDetailsHomeComponent } from "./details/application-details-home.component";
-import { ApplicationDetailsComponent } from "./details/application-details.component";
-import { ApplicationPackagesComponent } from "./details/application-packages.component";
-import { ApplicationPropertiesComponent } from "./details/application-properties.component";
-import { ApplicationErrorDisplayComponent } from "./errors/application-error-display.component";
-import { ApplicationHomeComponent } from "./home/application-home.component";
+import { ActivatePackageDialogComponent, ApplicationCreateDialogComponent, ApplicationEditDialogComponent,
+    DeleteApplicationDialogComponent,
+} from "./action";
+import { ApplicationListComponent, ApplicationPreviewComponent } from "./browse";
+import { ApplicationDefaultComponent, ApplicationDetailsComponent , ApplicationPackageTableComponent,
+    ApplicationPackagesComponent, ApplicationPropertiesComponent,
+} from "./details";
+import { ApplicationErrorDisplayComponent } from "./errors";
+import { ApplicationHomeComponent } from "./home";
 
-const components = [ActivatePackageDialogComponent, ApplicationCreateDialogComponent, ApplicationDetailsHomeComponent,
+const components = [ActivatePackageDialogComponent, ApplicationCreateDialogComponent, ApplicationDefaultComponent,
     ApplicationDetailsComponent, ApplicationEditDialogComponent, ApplicationErrorDisplayComponent,
-    ApplicationHomeComponent, ApplicationListComponent, ApplicationPreviewComponent, ApplicationPackagesComponent,
-    ApplicationPropertiesComponent, DeleteApplicationDialogComponent];
+    ApplicationHomeComponent, ApplicationListComponent, ApplicationPackagesComponent, ApplicationPackageTableComponent,
+    ApplicationPreviewComponent, ApplicationPropertiesComponent, DeleteApplicationDialogComponent];
 
 @NgModule({
     declarations: components,
