+++ app/components/application/browse/application-list.html
@@ -4,7 +4,7 @@
         </bex-quick-list-item-status>
 
         <h4 bex-quick-list-item-title>{{application.id}}</h4>
-        <p bex-quick-list-item-field>{{application.packages?.length}} packages</p>
+        <p bex-quick-list-item-field>{{application.packages?.size}} packages</p>
         <div bex-quick-list-item-preview>
             <bex-application-preview [application]="application"></bex-application-preview>
         </div>
