--- app/components/application/browse/preview/index.ts
@@ -1,1 +1,0 @@
-export * from "./application-preview.component";
