+++ app/components/application/details/application-packages.html
@@ -1,20 +1,23 @@
-<bex-table>
-    <bex-thead>
-        <bex-column>Version</bex-column>
-        <bex-column>State</bex-column>
-        <bex-column>Format</bex-column>
-        <bex-column>Last activation time</bex-column>
-    </bex-thead>
-    <bex-row *ngFor="let package of displayedPackages" [key]="package.version">
-        <bex-cell style="width: 25%" [value]="package.version"></bex-cell>
-        <bex-cell style="width: 25%" [value]="package.state"></bex-cell>
-        <bex-cell style="width: 25%" [value]="package.format"></bex-cell>
-        <bex-cell style="width: 25%" [value]="formatDate(package.lastActivationTime)"></bex-cell>
-    </bex-row>
-</bex-table>
+<bex-entity-details-list #listWrapper [showRefreshButton]="false" [showExpandButton]="false" [showFilterIcon]="false"
+    addButtonHoverText="Add package version" (onAddEntity)="list?.addPackage($event)" filterPlaceholder="Filter by package version">
+    <div bex-list-buttons>
+        <!--
+            multi select - "add" enabled and "delete" enabled if allow updates is true
+            single select
+                - delete: enabled if allow updates is true
+                - update: enabled if allow updates is true and package is active
+                - activate: enabled if package is pending
+        -->
+        <button md-mini-fab color="warn" title="Delete selected" (click)="list?.deleteSelectedItems()" [disabled]="!(list?.deleteItemEnabled | async)">
+            <i class="fa fa-trash spin-hover"></i>
+        </button>
+        <button md-mini-fab color="primary" title="Update package" (click)="list?.updatePackageVersion()" [disabled]="!(list?.editItemEnabled | async)">
+            <i class="fa fa-pencil-square-o spin-hover"></i>
+        </button>
+        <button md-mini-fab color="primary" title="Activate pending package" (click)="list?.activateActiveItem()" [disabled]="!(list?.activateItemEnabled | async)">
+            <i class="fa fa-toggle-on spin-hover"></i>
+        </button>
+    </div>
 
-<bex-no-item [itemCount]="displayedPackages.size" [filter]="filter">
-    <i icon class="fa fa-cogs"></i>
-    <span no-filter>This application contains no package versions</span>
-    <span with-filter>Current filter returned no matches</span>
-</bex-no-item>
+    <bex-application-package-table [application]="application" [filter]="listWrapper.filter" #list></bex-application-package-table>
+</bex-entity-details-list>
