+++ app/components/application/details/index.ts
@@ -1,0 +1,5 @@
+export * from "./application-default.component";
+export * from "./application-details.component";
+export * from "./application-package-table.component";
+export * from "./application-packages.component";
+export * from "./application-properties.component";
