+++ app/components/application/errors/index.ts
@@ -1,0 +1,1 @@
+export * from "./application-error-display.component";
