+++ app/components/application/home/index.ts
@@ -1,0 +1,1 @@
+export * from "./application-home.component";
