+++ app/components/job/details/job-details-home.component.ts
@@ -12,6 +12,6 @@ import { Component } from "@angular/core";
 
 export class JobDetailsHomeComponent {
     public static breadcrumb() {
-        return { name: "Job" };
+        return { name: "Jobs" };
     }
 }
