+++ app/components/task/details/task-resource-files.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnDestroy, ViewContainerRef } from "@angular/core";
+import { Component, Input, ViewContainerRef } from "@angular/core";
 
 import { ResourceFile, Task } from "app/models";
 
@@ -7,7 +7,7 @@ import { ResourceFile, Task } from "app/models";
     templateUrl: "task-resource-files.html",
 })
 
-export class TaskResourceFilesComponent implements OnDestroy {
+export class TaskResourceFilesComponent {
     @Input()
     public set task(task: Task) {
         this._task = task;
@@ -27,9 +27,5 @@ export class TaskResourceFilesComponent implements OnDestroy {
         if (this.task) {
             this.resourceFiles = this.task.resourceFiles || [];
         }
-    }
-
-    public ngOnDestroy() {
-        /* tab hide */
     }
 }
